+++ examples/wasm-app/README.md
@@ -1,3 +1,6 @@
+# THIS EXAMPLE DOES NOT WORK AT THE MOMENT
+Will be unmaintained until more of yage works/is implemented cleanly.
+
 # `wasm-app`
 
 Minimal example bootstrapped with [create-wasm-app](https://github.com/rustwasm/create-wasm-app).
