--- yage-app/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "yage-app"
-version = "0.0.1"
-authors = ["Benjamin Wasty <benny.wasty@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-glutin = "0.19.0"
