--- yage-app/README.md
@@ -1,3 +1,0 @@
-This crate should abstract over native window and event handling (glutin) and WebGL canvases in the browser. Potential inspiration: https://github.com/unrust/uni-app
-
-TODO: better name?
