+++ yage-gltf/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "yage-gltf"
 version = "0.0.1"
-authors = ["Benjamin Wasty <benny.wasty@gmail.com>"]
+authors = ["Benjamin Wasty <benjamin.wasty@cginternals.com>", "Stefan Buschmann <stefan.buschmann@cginternals.com>"]
 edition = "2018"
 
 [dependencies]
