+++ yage-glutin/Cargo.toml
@@ -1,10 +1,8 @@
 [package]
-name = "native-glutin"
+name = "yage-glutin"
 version = "0.0.1"
 authors = ["Benjamin Wasty <benjamin.wasty@cginternals.com>", "Stefan Buschmann <stefan.buschmann@cginternals.com>"]
 edition = "2018"
 
 [dependencies]
 glutin = "0.19.0"
-gl = "0.11.0"
-yage = { path = "../../" }
