+++ yage-glutin/README.md
@@ -1,0 +1,1 @@
+This crate should handle window, context, events loop (most of what is now in `examples/native-glutin/main.rs`)
