+++ yage-web/README.md
@@ -1,0 +1,1 @@
+This crate should handle the WebGL context and be the web counterpart of yage-glutin.
